+++ src/lib/api/AuthClient/index.ts
@@ -1,0 +1,2 @@
+export * from "./RestSessionAuthClient";
+export * from "./types";