+++ src/lib/api/AuthClient/types.ts
@@ -1,0 +1,6 @@
+export type RestSessionAuthClientLogoutOptions = {
+  accessToken?: string;
+  refreshToken?: string;
+};
+
+export type AuthClientLogoutOptions = RestSessionAuthClientLogoutOptions;