+++ src/lib/api/RestAuthApiClient/constants.ts
@@ -1,0 +1,6 @@
+export const DEFAULT_REST_AUTH_API_CLIENT_ENDPOINTS = {
+  forgotPassword: "auth/password/forgot",
+  resetPassword: "auth/password/reset",
+  resendConfirmEmail: "auth/email/confirm/resend",
+  confirmEmail: "auth/email/confirm",
+} as const;