+++ src/lib/api/SupabaseDataClient/index.ts
@@ -1,0 +1,2 @@
+export * from "./SupabaseDataClient";
+export * from "./types";